Improve DateTime resolution for job arguments
At job execution if a DateTime argument was included the DateTime object does not include milliseconds. The issue is caused at job creation when converting arguments to their string representation. The default conversion does not provide milliseconds. This is backwards compatible with existing data since a new DateTime object can be instantiated with both the old format and this new format, example:

old: "05/30/2014 07:27:34"
new: "05/30/2014 07:27:34.2791"

If there is a better approach I'd be happy to make the change.