Add param to Job ctor to indicate missing method
Starts to resolve #157

InvocationData almost gives us what we need when it throws an InvalidOperationException on deserialising if the method is null. The actual exception message thrown is almost exactly what should be shown in the Dashboard.

This commit adds a boolean parameter to the internal Job constructor, so that it can be told that the method is missing. A boolean property was also added to Job, so that this can be examined at a later stage to determine whether or not the job should be processed, and whether we need to display a message with the missing method details in it.

Also fixed all the subsequently broken calls in all the tests to have the new fourth parameter.