Remove the X axis from the realtime graph
It caused "Cannot read property 'removeChild' of null" errors appeared
in Chrome and IE when opening dashboard for the first time.
I've tried to use Rickshaw.Graph.Axis.X, but it gives the same error. So
I decided to remove it, as it don't add a lot of value – hover tooltips
are working fine.
Closes #137.