Fix very rare resource leaks detected by Coverity Scan
There are no leaks in this code currently, but let's guard ourselves
against them in future. And to calm down Coverity Scan :-)

DbConnection was leaking in SqlServerJobQueue, when a call to
BeginTransaction was failed with an exception, when MARS is used.
Parallel transactions aren't possible here anyway.
MsmqTransaction was leaking, when Receive method call throws exception,
other than MessageQueueException. This exception may appear only due to
programming errors in future.