Change order of the QueueAttribute state filter to be always the last one
It's better to make it run as an apply state filter, but this is a
breaking change. This filter always should be called last, especially if
previous filters are suspected in changing candidate state to the
enqueued one, to not to ignore the queue setting. This behavior leads to
bug, when continuation is created when antecedent job is already
finished, and our continuation is enqueued to the default queue,
regardless of the specified one.