Optimize the distributed lock acquire algorithm appeared in 1.6.15
Previous implementation may cause much more queries, when acquiring a
lock on a blocked resource, when high value is passed in the timeout
argument. We'll pass second-based timeout instead to the sp_getapplock
procedure instead – it's not too high to cause thread pool starvation,
and large enough to significantly reduce the number of queries.