Decrease the number of roundtrips when committing a transaction
One can now optionally use
SqlServerStorageOptions.BatchCommandMaxTimeout property to enable
command batching inside transactions. When set, almost all the commands
will be flushed in one network call due to using SqlCommandSet class on
.NET Framework.
This greatly reduces the time required for a transaction to commit,
especially when a lot commands or slow network is involved, resulting to
much fewer command timeouts for transaction itself, as well as for other
transaction affected by this one.