Use custom lock in SqlConnection.SetRangeInHash to prevent deadlocks
There are a lot of indexes in the Hash table, and the clustered one is
simply based on an incremental id. Without using a lock as in
transactions, sooner or later this method will cause a deadlock. Without
using a lock as in transactions, sooner or later this method will cause
a deadlock.
Fixes #1124