Force DequeueUsingSlidingInvisibilityTimeout to use page locks
This is related to so called "ghost" rows in SQL Server. By default, it
collects them asynchronously, after some threshold reached, but when
PAGLOCK is specified they are removed immediately. The absence of ghost
rows allows to significantly speed up scan operations and reduce the CPU
consumption, because less rows are traversed and less locks are taken.