Adding GetFirstByLowestScoreFromSet overload (#1251)
* scaffolding overload of grabbing next N jobs from set

* implementing overload, parameterizing TOP statement w @max param

* adding coverage of our new GetLowestScoreFromSet overloaded method

* fixing encoding of copyright char

* Remove new GetFirstByLowerScoreFromSet overload from the interface

By introducing the new overload only in an abstract class, and making it virtual, we are removing breaking changes, making the implementation optional. So all the current storages will be compatible with these changes.

* Rename "max" argument to "count" in the new overload

Just to use the same words as in other places like IMonitoringApi.

* Change tests to check actual parameter that caused an exception

Since ArgumentException can be originated from different parameters, we also need to check the actual parameter that caused an exception.

* Change return value of the new overload to List<string>

The result is an ordered collection, and that order is based on `score` rather than on the value itself. Since HashSet can't preserve that type of ordering, we are using List instead.