Add more guards to InvocationData's logic to not to depend on serialization helper
The test that's implemented in this commit passes just fine even without additional logic presented here with the new serialization helper class. However if we change any of the setting we may end up in an unwanted errors. So this logic is just to reassure we don't get deserialization errors even if something is changed.