Ensure RecurringJobScheduler doesn't go into infinite loop
There's a chance that some old servers (pre-1.7.0) are still running in our environment, and since they don't know anything about scores in the recurring job sets, they don't update them.  Without this change it is possible for the new server to pick the same job again and again without scheduling it.