Ensure we can deserialize EnqueuedState even if it doesn't contain queue
There was a change in earlier betas that omits the "Queue" property from serialized payloads. I've rolled back that change, because there's no easy change to make rolling upgrades work in this case. However it's possible that early users still have data serialized this way, and we should handle it.