Make TaskScheduler.Default the default scheduler to avoid breaking changes
There may be job methods, where Task.Result or Task.Wait is used, and in previous versions this usage didn't result to deadlocks. To not to cause deadlocks to appear with the same settings, we are switching to the default task scheduler that can handle them automatically by adding more threads. However this default should be changed in 2.0.0.