Make background job creation resilient to transient exceptions
CoreBackgroundJobFactory class now supports idempotent retries for background job creation logic that don't lead to multiple similar jobs even if a timeout exception is thrown when attempt was actually succeeded (i.e. TransactionInDoubt state in ADO.NET and so on).

The new behavior can be enabled by setting the BackgroundJobClient.RetryAttempts to a positive value, and it will be enabled by default in version 2.0.