_options.NonBlockingFetchSql added
I am huge fan of Hangfire I am using it for multiple environments however I found that the most resource hogger for our sql server is the sp_getapplock. Please not this server is a huge report server with many data load (importing data from the cloud), and we get the the most wait time for sp_getapplock.