Add UseFineGrainedLocks option for SQL Server storage
It will prevent theoretical deadlocks caused by the fact that SQL Server's range lock holds not only for the keys of the target range, but also b-tree node of the next key is locked too. I didn't get a confirmation of this behavior yet, but we'll have a switch that's ready to be used.