Don't let RecurringJobScheduler to stall the pipeline when extension filters throw an exception
As with one of previous commits, before Hangfire 1.6.20 background state changer itself was handling all the exceptions occurred in state changing filters. But in that version exception handling logic was move to the Worker class to make it more flexible. But it wasn't moved to delayed and recurring job schedulers as well.