Remove any mentions of code coverage tools to avoid confusion
Related to the recent CodeCov breach. Actually Codecov wasn't used since migration to the new project format in version 1.7.0, because I was lazy enough to implement such a support.
Coverity scan is also being removed due to its complexity and false positives (but I may be wrong, because didn't check it for 2 or 3 years).