Add the ScheduledState.Queue property to speed up delayed jobs
The value of this property is used by the delayed job scheduler component to enqueue such jobs directly to the specified queue bypassing state transition. This significantly improves performance of delayed jobs, and workers are able to process them directly with almost the same speed as enqueued ones.

However such a queue name will not be considered during regular state transitions, because Hangfire does not read metadata of a previous state during such state changes. We change the semantics of the state, and scheduled state with the Queue property set changes how job in this state will be processed – it will be placed into the queue directly, the same way how Enqueued state works.

As another side effect – with these changes delayed job scheduler does not require to have access to background job methods and can schedule jobs of different code bases to different queues.