Don't use readcommittedlock when not required
This table hint is now used only in GetJobData and GetStateData to get linearizable reads when using external queue implementation (like MSMQ or RabbitMQ) and read committed snapshot isolation option is turned on in SQL Server (as by default in SQL Azure). It was set for other queries just in case, but it's not required, and SQL Azure can perform much better when this hint is not used.