Don't display queue name when it's default
Since EnqueuedState.Queue property is now used to override the Job.Queue property value, we shouldn't display the default queue to avoid confusion when looking at the job details page – e.g. when Job.Queue is set to "critical", and EnqueuedState.Queue is set to "default", in this case job will be scheduled to the "critical" queue, not the "default" as displayed in old history renderer.