Deprecate methods with optional parameters in RecurringJob
It's not possible to alter these methods without introducing breaking changes, and we are required to make some changes due to the new role of the "queue" aspect introduced in the previous commits. So confusion will be added with this and later commits, but it's a necessary evil to make the transition with no breaking changes.