Avoid any blocked rows when removing inactive servers
Deadlock is reported between "Update Heartbeat" and "Remove Inactive Servers" queries. This was really strange deadlock, because heartbeat query modifies only a single row with initial "update" lock. But anyway it's not required for this query to remove locked rows, so it's not a problem to include the "readpast" hint to avoid any deadlocks.