Deprecate UsePageLocksOnDequeue, now it does nothing
This option sometimes increases job throughput with on-premise SQL Server, SSD disk and when multiple queues are used, because of instant phantom row deletion since page lock in X mode is applied. However concurrent insert to the job queue in client-managed transaction causes really long delays, making all the improvements of this option to disappear.