Avoid throwing OCE in while(true) loops
When hosted in IIS, when deploying new app version, when using autostart providers (probably) and when appdomain unload was delayed (probably), OperationCancelledException thrown in the while(true) or while loop with some external condition, for some reason doesn't lead to loop exit, and leads to new iteration instead without any delay. This behavior was observed when BackgroundExecution class was written and looked really weird, and it's better to avoid calling ThrowIfCancellationRequested in such loops to avoid bad things.

Relates to #1962