Better protection from wait failures in executions
We use waits to protect the code from looping again and again on exceptions, because this behavior can possibly result to denial of service, so we rely on the fact that every wait should be succeeded. And now execution is able to shutdown itself with a fatal exception when wait was failed.