Bump Microsoft.Owin package to version 4.2.2
This package is included in the Hangfire.Core one by meging and internalizing via the ILRepack program, however only very limited subset of features is used there, excluding cookies parsing for example. So despite of there are reported vulnerabilities (GHSA-3rq8-h3gj-r5c6 and GHSA-hxrm-9w7p-39cc) in the Microsoft.Owin package of version 3.0.1, this is not a security-related change, since we don't use any of the affected features and don't expose them to public via internalizing.