Exclude Newtonsoft.Json from NuGet Audit in the Hangfire.Core project
The Newtonsoft.Json package is defined a dependency only and is not merged with the Hangfire.Core.dll assembly. Versions defined in the project file and in the nuspec file serve as a minimal required version, and Hangfire.Core package can be used with any higher version installed in the target application by explicitly adding reference to its project file.
Vulnerabilities are detected in each version of Newtonsoft.Json package lower than 13.0.1, and we can't always bump Hangfire.Core dependency to a higher version without introducing breaking changes to a lot of users and creating a risk of dependency hell, where different packages included in a target application require different versions that can't be installed together.